metrics/fastapi-otel/python/requirements.txt
annotated-doc==0.0.4 ; python_version >= "3.10" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==4.12.0 ; python_version >= "3.10" and python_version < "4.0"
asgiref==3.11.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2026.1.4 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.4.4 ; python_version >= "3.10" and python_version < "4.0"
click==8.3.1 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
exceptiongroup==1.3.1 ; python_version == "3.10"
fastapi==0.128.0 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.72.0 ; python_version >= "3.10" and python_version < "4.0"
grpcio==1.76.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.11 ; python_version >= "3.10" and python_version < "4.0"
importlib-metadata==8.7.1 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-api==1.39.1 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-exporter-otlp-proto-common==1.39.1 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-exporter-otlp-proto-grpc==1.39.1 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-exporter-otlp-proto-http==1.39.1 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-exporter-otlp==1.39.1 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-exporter-prometheus==0.60b1 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-asgi==0.60b1 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-fastapi==0.60b1 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-requests==0.60b1 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation==0.60b1 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-proto==1.39.1 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-sdk==1.39.1 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-semantic-conventions==0.60b1 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-util-http==0.60b1 ; python_version >= "3.10" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.10" and python_version < "4.0"
prometheus-client==0.23.1 ; python_version >= "3.10" and python_version < "4.0"
protobuf==6.33.2 ; python_version >= "3.10" and python_version < "4.0"
pydantic-core==2.41.5 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.12.5 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.50.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.10" and python_version < "4.0"
typing-inspection==0.4.2 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.6.2 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.40.0 ; python_version >= "3.10" and python_version < "4.0"
wrapt==1.17.3 ; python_version >= "3.10" and python_version < "4.0"
zipp==3.23.0 ; python_version >= "3.10" and python_version < "4.0"
